Bring back custom resolver, update dependencies
