+++ addon/package.json
@@ -92,14 +92,9 @@
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.2.1",
     "expect-type": "^0.15.0",
-    "fs-extra": "^11.1.0",
-    "latest-version": "^5.0.0",
     "loader.js": "^4.7.0",
-    "pretender": "^3.4.7",
     "prettier": "^2.8.8",
     "qunit": "^2.19.4",
-    "release-it": "~15.6.0",
-    "release-it-lerna-changelog": "^5.0.0",
     "rimraf": "^5.0.1",
     "typescript": "^5.1.3",
     "webpack": "^5.86.0"
@@ -111,39 +106,11 @@
     "access": "public",
     "registry": "https://registry.npmjs.org/"
   },
-  "changelog": {
-    "repo": "emberjs/ember-test-helpers",
-    "labels": {
-      "breaking": ":boom: Breaking Change",
-      "enhancement": ":rocket: Enhancement",
-      "bug": ":bug: Bug Fix",
-      "documentation": ":memo: Documentation",
-      "internal": ":house: Internal"
-    }
-  },
   "ember": {
     "edition": "octane"
   },
   "ember-addon": {
     "configPath": "tests/dummy/config"
-  },
-  "release-it": {
-    "scripts": {
-      "after:bump": "pnpm docs"
-    },
-    "plugins": {
-      "release-it-lerna-changelog": {
-        "infile": "CHANGELOG.md",
-        "launchEditor": true
-      }
-    },
-    "git": {
-      "tagName": "v${version}"
-    },
-    "github": {
-      "release": true,
-      "tokenRef": "GITHUB_AUTH"
-    }
   },
   "volta": {
     "extends": "../package.json"
