+++ test-app/app/app.js
@@ -1,5 +1,5 @@
 import Application from '@ember/application';
-import Resolver from 'ember-resolver';
+import Resolver from './resolver';
 import loadInitializers from 'ember-load-initializers';
 import config from 'test-app/config/environment';
