+++ test-app/package.json
@@ -31,6 +31,7 @@
     "@ember/optional-features": "^2.0.0",
     "@ember/string": "^3.0.1",
     "@ember/test-helpers": "workspace:*",
+    "pretender": "^3.4.7",
     "@embroider/test-setup": "^4.0.0",
     "@glimmer/component": "^1.1.2",
     "@glimmer/tracking": "^1.1.2",
@@ -50,7 +51,7 @@
     "ember-load-initializers": "^2.1.2",
     "ember-modifier": "^4.1.0",
     "ember-page-title": "^7.0.0",
-    "ember-qunit": "^6.2.0",
+    "ember-qunit": "^8.1.0",
     "ember-resolver": "^10.0.0",
     "ember-source": "~4.12.0",
     "ember-source-channel-url": "^3.0.0",
@@ -66,8 +67,8 @@
     "latest-version": "^5.0.0",
     "loader.js": "^4.7.0",
     "prettier": "^2.8.7",
-    "qunit": "^2.19.4",
-    "qunit-dom": "^2.0.0",
+    "qunit": "^2.21.1",
+    "qunit-dom": "^3.2.0",
     "stylelint": "^15.4.0",
     "stylelint-config-standard": "^32.0.0",
     "stylelint-prettier": "^3.0.0",
